Remove SwiftyMarkdown
