+++ Scripts/updateDependency.sh
@@ -25,8 +25,6 @@ if [[ $dependency_directory == *"Nuke"* ]]; then
 	dependency_url="git@github.com:kean/Nuke.git"
 elif [[ $dependency_directory == *"SwiftyGif"* ]]; then
 	dependency_url="git@github.com:kirualex/SwiftyGif.git"
-elif [[ $dependency_directory == *"SwiftyMarkdown"* ]]; then
-    dependency_url="git@github.com:GetStream/SwiftyMarkdown.git"
 elif [[ $dependency_directory == *"DifferenceKit"* ]]; then
 	dependency_url="git@github.com:ra1028/DifferenceKit.git"
 else
@@ -69,11 +67,4 @@ rm -rf $dependency_directory
 if [[ $dependency_directory == *"DifferenceKit"* ]]; then
     # We currently use customized UIKit extensions in Utils/DifferenceKit+Stream.swift
     rm $output_directory/Extensions/UIKitExtension.swift
-fi
-
-if [[ $dependency_directory == *"SwiftyMarkdown"* ]]; then
-    # We currently use customized version of SwiftyMarkdown
-	git restore $output_directory/SwiftyMarkdown/PerformanceLog.swift || true
-    git restore $output_directory/SwiftyMarkdown/SwiftyLineProcessor.swift || true
-    git restore $output_directory/SwiftyMarkdown/SwiftyTokeniser.swift || true
 fi
